backend/admin-auth/requirements.txt
pyjwt>=2.8.0
bcrypt>=4.1.0
psycopg2-binary>=2.9.9
